fix(eip8130): keep balance-monitor tests colocated; guard no_std phase-status drop
Address review feedback on the receipt PR:

- Revert the balance-monitor unit tests back to a colocated `#[cfg(test)] mod
  tests` block in `monitor.rs`, per the project convention against standalone
  `tests.rs` modules. This drops the unrelated test relocation entirely.
- Add a `debug_assert!` in the `no_std` `Eip8130PhaseStatuses::set` so a future
  caller that drops non-empty statuses (which `take` could not recover) fails
  loudly instead of silently; EIP-8130 execution is `std`-gated, so statuses are
  always empty here today.
