+++ crates/utilities/balance-monitor/src/lib.rs
@@ -9,6 +9,3 @@
 
 mod monitor;
 pub use monitor::BalanceMonitorLayer;
-
-#[cfg(test)]
-mod tests;
